+++ src/libs/decoders/wav.c
@@ -140,7 +140,7 @@ static Uint32 WAV_read(Sound_Sample *sample)
     const drwav_uint64 frames_read = drwav_read_pcm_frames_s16(dr,
                                          internal->buffer_size / (dr->channels * sizeof(drwav_int16)),
                                          (drwav_int16 *) internal->buffer);
-    return frames_read * dr->channels * sizeof (drwav_int16);
+    return (Uint32)(frames_read * dr->channels * sizeof (drwav_int16));
 } /* WAV_read */
